+++ src/cmp/Layout/cmp/Panel/cmp/Symbols/Sym/cmp/Label/index.jsx
@@ -21,7 +21,7 @@ const Label = ({sym, checked, onClick}) => {
     return <div className={localCasses.SectionLabel}>
         <div>
             <Box className={classes.Box}>
-                {Boolean(symbols.length > 1) && <input type="checkbox" checked={checked} onClick={onClick}/>}
+                {Boolean(symbols.length > 1) && <input type="checkbox" checked={checked} onChange={onClick}/>}
                 <strong>Label: </strong><input type="text" onInput={updateLabel} size="small" label="Symbol label"  value={label} />
             </Box>
         </div>